Set RPATH linker flag to $ORIGIN.
This allows to add shared libraries in the same directory as gmic-qt on
the Linux platform (allowing a behavior similar to Windows which
searches for libraries in the executable folder).
Note that I set both CMAKE_EXE_LINKER_FLAGS and CMAKE_INSTALL_RPATH. The
later should normally be enough, but updates only the plug-in during
installation (which seems to be a bogus step in this repo).

This is a first step to try and get the G'Mic plug-in to run correctly
on all platforms (e.g. flatpak).