Make the Apply button work with the filter preview
When the Apply button is clicked in the G'MIC-Qt user interface the
active layer will be replaced with the first preview image.
This allows users to "layer" multiple G'MIC effects before returning
the final image to the 8bf plug-in.

Note that only the most recently applied effect will be used when the
8bf plug-in is invoked through the "Last Filter" or "Repeat Filter"
commands.