Fix gimp3 building
Each of the active layer ID comparisons was failing to compile for me:

    error: ordered comparison of pointer with integer zero
    (‘GimpLayer*’ and ‘int’)

If the comparison were allowed, it’d always be true, pointers being
unsigned. But the compiler isn’t liking it.

Comparing with _GIMP_NULL_LAYER, I presume that the `activeLayerID >= 0`
comparisons need to become `activeLayerID != NULL`.

No idea if `activeLayerID != _GIMP_NULL_LAYER` would be acceptable or
not; dunno if old GIMP did anything with -2, -3, &c. or if -1 was the
only special value. Without that special knowledge, I figure that the
macro I’ve added is probably safer.

I emphasise that as I submit this patch I do not know if it’s correct.
All I do know is that with the patch it *compiles*, and trivial
functionality seems to work.