Add color management for the G'MIC-Qt preview
This improves the appearance of the preview image for all image modes.
Color management will only be used if the 8bf plug-in host provides the
image color profile to its 8bf plug-ins.

G'MIC-Qt will attempt to retrieve the color profile for the display
that G'MIC-Qt is using when the first call to applyColorProfile is
made, if that is unsuccessful the color profile of the primary display
will be used.
The display color profile will not be refreshed if the user moves the
G'MIC-Qt UI to a different display within the same session.

Depending on the image, the 32-bit float mode may still have color
differences versus Photoshop's native rendering.
I am not sure if my code is missing a step in the color correction
process, or if it is a difference in behavior between Little CMS and
Adobe's color management engine.